+++ src/Enums/DeviceOrientations.js
@@ -1,0 +1,7 @@
+/**
+ * DeviceOrientations
+ */
+export default {
+  PORTRAIT: 1,
+  LANDSCAPE: 2,
+};