+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased](https://github.com/Teknasyon-Teknoloji/deepwall-react-native-sdk/compare/2.9.5...main)
+### Changed
+- On Android, `deepwall-core` version upgraded to version `2.6.2`
+- On android fixed extraData error when paywall opens
+- On iOS, `deepwall-core` version upgraded to version `2.4.2`
+
+### Added
+- On android added orientation parameter to `requestPaywall` method
 
 
 ---